Refactor:토큰 확인 로직 utils 함수로 생성
